Create PixelFormatRGBMask32 class
